re-move metabot stuff
