+++ frontend/src/metabase/admin/ai/CursorInstallLink.tsx
