+++ frontend/src/metabase/admin/ai/MCPServerUrlSection.tsx
