+++ frontend/src/metabase/admin/ai/McpAppsSettings.unit.spec.tsx
