+++ frontend/src/metabase/admin/ai/MetabotAdminLayout.tsx
@@ -1,5 +1,5 @@
 import ErrorBoundary from "metabase/ErrorBoundary";
-import { AdminSettingsLayout } from "metabase/common/components/AdminLayout/AdminSettingsLayout";
+import { AdminSettingsLayout } from "metabase/admin/components/AdminLayout/AdminSettingsLayout";
 import { Box } from "metabase/ui";
 
 import { MetabotNavPane } from "./MetabotNavPane";
