+++ frontend/src/metabase/admin/ai/MetabotAdminSuggestedPrompts.tsx
