+++ frontend/src/metabase/admin/ai/MetabotAdminSuggestedPrompts.unit.spec.tsx
