+++ frontend/src/metabase/admin/ai/MetabotNavPane.tsx
