+++ frontend/src/metabase/admin/ai/MetabotNavPane.unit.spec.tsx
