+++ frontend/src/metabase/admin/ai/MetabotSettingsPanel.unit.spec.tsx
