+++ frontend/src/metabase/admin/ai/MetabotSetup.unit.spec.tsx
