+++ frontend/src/metabase/admin/ai/test-utils.ts
