+++ frontend/src/metabase/metabot/components/MetabotManagedProviderLimit.tsx
@@ -2,6 +2,7 @@ import { useDisclosure } from "@mantine/hooks";
 import { useCallback } from "react";
 import { t } from "ttag";
 
+import { MetabotSetupInner } from "metabase/admin/ai/MetabotSetup";
 import { ExternalLink } from "metabase/common/components/ExternalLink";
 import { useStoreUrl } from "metabase/common/hooks";
 import { useDispatch } from "metabase/redux";
@@ -14,8 +15,6 @@ import {
   Modal,
   Text,
 } from "metabase/ui";
-
-import { MetabotSetupInner } from "./MetabotAdmin/MetabotSetup";
 
 const METABOT_MANAGED_PROVIDER_LIMIT_TOAST_ID =
   "metabot-managed-provider-limit";
